redux-thunk added, components refactored

+++ src/api/api.js
@@ -12,14 +12,13 @@ export const usersAPI = {
         return instance.get(`users?page=${currentPage}`).then(response => response.data)
     }
 }
-export const followUnfollowAPI = {
-    follow(id) {
-        return instance.post(`follow/${id}`).then(response => response.data)
-    },
-    unfollow(id) {
-        return instance.delete(`follow/${id}`).then(response => response.data)
+export const followUnfollowAPI = (id, param) => {
+       if (param === "post") {
+           return instance.post(`follow/${id}`).then(response => response.data)
+       } else {
+           return instance.delete(`follow/${id}`).then(response => response.data)
+       }
     }
-}
 
 export const profileAPI = {
     setProfile(userId) {
